move named_type.hpp into fwdpp/util
