+++ fwdpp/util/Makefile.am
@@ -1,6 +1,7 @@
 pkgincludedir=$(prefix)/include/fwdpp/util
 
 pkginclude_HEADERS=abstract_cloneable.hpp \
-				   enum_bitflags.hpp
+				   enum_bitflags.hpp \
+				   named_type.hpp
